v0.2.0
Now supports python 2 as well as python 3

wildcarded_ngrams() method was replaced with more robust
ngrams_by_pattern() method which expects a set of indices in the
provided ngram which are placeholders and can be replaced by any element

Profiling proved that the set of descendents in each node was not
increasing efficiency and so was removed.

Added unit tests file.